moved binary_exists into unit_tests.rs
